cheri: Use uintptr_t in migrate hooks
The callback data for the page migration hooks must be uintptr_t.
Adjust function prototypes etc. accordingly.

+++ mm/compaction.c
@@ -1794,7 +1794,7 @@ static void isolate_freepages(struct compact_control *cc)
  * This is a migrate-callback that "allocates" freepages by taking pages
  * from the isolated freelists in the block we are migrating to.
  */
-static struct folio *compaction_alloc_noprof(struct folio *src, unsigned long data)
+static struct folio *compaction_alloc_noprof(struct folio *src, uintptr_t data)
 {
 	struct compact_control *cc = (struct compact_control *)data;
 	struct folio *dst;
@@ -1842,7 +1842,7 @@ again:
 	return page_rmappable_folio(&dst->page);
 }
 
-static struct folio *compaction_alloc(struct folio *src, unsigned long data)
+static struct folio *compaction_alloc(struct folio *src, uintptr_t data)
 {
 	return alloc_hooks(compaction_alloc_noprof(src, data));
 }
@@ -1852,7 +1852,7 @@ static struct folio *compaction_alloc(struct folio *src, unsigned long data)
  * freelist.  All pages on the freelist are from the same zone, so there is no
  * special handling needed for NUMA.
  */
-static void compaction_free(struct folio *dst, unsigned long data)
+static void compaction_free(struct folio *dst, uintptr_t data)
 {
 	struct compact_control *cc = (struct compact_control *)data;
 	int order = folio_order(dst);
